requirements.txt
crawl4ai
google-cloud-firestore
tenacity
google-cloud-storage
google-cloud